requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-formrecognizer==3.3.3
azure-core==1.32.0
boto3==1.36.10
bs4==0.0.2
fastapi==0.115.8
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.28.1
streamlit==1.41.1
uvicorn==0.34.0
webdriver-manager==4.0.2